grayscale and feature detection
This is a package with which you can see the grayscale of your raw images as well as features in them (usual edges and corners). Make sure that the raw images are "rgb8" encoded and dimensions of your images are 360*640, if you want to directly plug and play (although the simple code will allow you to tweak things according to your convenience).